requirements.txt
streamlit
sounddevice
numpy
scipy
openai
langgraph
langchain-core
langchain-openai
reportlab
pillow
python-dotenv
